Use single-stage Docker build with multiple binaries
This commit changes the Dockerfile to have just one stage. We
still download mongod and include it in the container image,
but the ENTRYPOINT is just the Go binary. The k8s config has
also been updated to reference the single image and override
the command and arguments for the mongod container.